Started working on the CircleGeometryTool and fixed the bug that prevented SFCircleShapes from being updated via editor tools.
